refactor(src): separating main code from types
